// src/common/enums/itemCategory.ts
export enum ItemCategory {
  VISOR = 'visor',
  SUIT = 'suit',
  BEAM = 'beam',
  MORPH_BALL = 'morph_ball',
  MOVEMENT = 'movement',
  MISSILE = 'missile',
  CHARGE_COMBO = 'charge_combo',
  ENERGY_TANK = 'energy_tank',
  CHOZO_ARTIFACT = 'chozo_artifact',
  MISSILE_RELATED = 'missile_related',
  MORPH_BALL_RELATED = 'morph_ball_related',
  BEAM_RELATED = 'beam_related',
  LIFE_SUPPORT = 'life_support',
  EXPANSION = 'expansion'
}